Split test constants
