+++ src/test/routes/token.constants.json
@@ -1,0 +1,27 @@
+{
+  "initialUser": {
+      "userName": "abrden",
+      "password": "1234",
+      "firstName": "Agustina",
+      "lastName": "Barbetta",
+      "country": "Argentina",
+      "email": "agustina.barbetta@gmail.com",
+      "birthdate": "120894",
+      "images": [
+        "lala",
+        "lolo",
+        "lili"
+      ]
+  },
+  "invalidCredentials": {
+    "userName": "tano_villano",
+    "password": "botox8"
+  },
+  "tokenGeneration": {
+    "userName": "abrden",
+    "password": "1234"
+  },
+  "tokenGenerationWithMissingAttributes": {
+    "userName": "abrden"
+  }
+}
