+++ src/test/routes/user.constants.json
@@ -60,16 +60,5 @@
     "userName": "abrden",
     "password": "4321",
     "email": "aibarbetta@fi.uba.ar"
-  },
-  "invalidCredentials": {
-    "userName": "tano_villano",
-    "password": "botox8"
-  },
-  "tokenGeneration": {
-    "userName": "abrden",
-    "password": "1234"
-  },
-  "tokenGenerationWithMissingAttributes": {
-    "userName": "abrden"
   }
 }
